+++ tests/Moq.Tests/ProtectedAsMockFixture.cs
@@ -167,9 +167,10 @@ namespace Moq.Tests
 			Assert.Equal(42, actual);
 		}
 
-		[Fact] void SetupGet_can_setup_virtual_property()
+		[Fact]
+		public void SetupGet_can_setup_virtual_property()
 		{
-			this.protectedMock.SetupGet(m => m.Virtual).Returns(42);
+			this.protectedMock.SetupGet(m => m.VirtualGet).Returns(42);
 
 			var actual = mock.Object.GetVirtual();
 
@@ -237,6 +238,96 @@ namespace Moq.Tests
 			this.mock.Object.DoSomething();
 
 			Assert.IsType<InvalidOperationException>(exception);
+		}
+
+		[Fact]
+		public void SetUpSet_should_setup_setters()
+		{
+			this.protectedMock.SetupSet(fish => fish.ReadWritePropertyImpl = 999).Throws(ExpectedException.Instance);
+
+			mock.Object.ReadWriteProperty = 123;
+
+			Assert.Throws<ExpectedException>(() => mock.Object.ReadWriteProperty = 999);
+		}
+
+		[Fact]
+		public void SetUpSet_should_setup_setters_with_property_type()
+		{
+			int value = 0;
+			this.protectedMock.SetupSet<int>(fish => fish.ReadWritePropertyImpl = 999).Callback(i => value = i);
+
+			mock.Object.ReadWriteProperty = 123;
+			Assert.Equal(0, value);
+
+			mock.Object.ReadWriteProperty = 999;
+			Assert.Equal(999, value);
+		}
+
+		[Fact]
+		public void SetUpSet_should_work_recursively()
+		{
+			this.protectedMock.SetupSet(f => f.Nested.Value = 999).Throws(ExpectedException.Instance);
+
+			mock.Object.GetNested().Value = 1;
+			
+			Assert.Throws<ExpectedException>(() => mock.Object.GetNested().Value = 999);
+		}
+
+		[Fact]
+		public void SetUpSet_Should_Work_With_Indexers()
+		{
+			this.protectedMock.SetupSet(
+				o => o[
+					It.IsInRange(0, 5, Range.Inclusive),
+					It.IsIn("Bad", "JustAsBad")
+				] = It.Is<int>(i => i > 10)
+			).Throws(ExpectedException.Instance);
+
+			mock.Object.SetMultipleIndexer(1, "Ok", 999);
+
+			Assert.Throws<ExpectedException>(() => mock.Object.SetMultipleIndexer(1, "Bad", 999));
+		}
+
+		[Fact]
+		public void SetupSet_can_setup_virtual_property()
+		{
+			this.protectedMock.SetupSet(m => m.VirtualSet = 999).Throws(new ExpectedException());
+
+			mock.Object.SetVirtual(123);
+			Assert.Throws<ExpectedException>(() => mock.Object.SetVirtual(999));
+		}
+
+		[Fact]
+		public void VerifySet_Should_Work()
+		{
+			void VerifySet(Times? times = null,string failMessage = null)
+			{
+				this.protectedMock.VerifySet(
+				o => o[
+					It.IsInRange(0, 5, Moq.Range.Inclusive),
+					It.IsIn("Bad", "JustAsBad")
+				] = It.Is<int>(i => i > 10),
+				times,
+				failMessage
+				);
+			}
+			VerifySet(Times.Never());
+
+			mock.Object.SetMultipleIndexer(1, "Ok", 1);
+			VerifySet(Times.Never());
+
+			Assert.Throws<MockException>(() => VerifySet()); // AtLeastOnce
+
+			mock.Object.SetMultipleIndexer(1, "Bad", 999);
+			VerifySet(); // AtLeastOnce
+			
+			mock.Object.SetMultipleIndexer(1, "JustAsBad", 12);
+			VerifySet(Times.Exactly(2));
+
+			Assert.Throws<MockException>(() => VerifySet(Times.AtMostOnce()));
+
+			var mockException = Assert.Throws<MockException>(() => VerifySet(Times.AtMostOnce(),"custom fail message"));
+			Assert.StartsWith("custom fail message", mockException.Message);
 		}
 
 		[Fact]
@@ -326,7 +417,7 @@ namespace Moq.Tests
 
 		public interface INested
 		{
-			int Value { get; }
+			int Value { get; set; }
 			int Method(int value);
 		}
 
@@ -376,23 +467,49 @@ namespace Moq.Tests
 				return Nested;
 			}
 
-			private int virtualProperty;
-			public virtual int Virtual
+			protected abstract int this[int i, string s] { get; set; }
+
+			public void SetMultipleIndexer(int index, string sIndex, int value)
+			{
+				this[index, sIndex] = value;
+			}
+
+			private int _virtualSet;
+			public virtual int VirtualSet
+			{
+				get
+				{
+					return _virtualSet;
+				}
+				protected set
+				{
+					_virtualSet = value;
+				}
+
+			}
+
+			public void SetVirtual(int value)
+			{
+				VirtualSet = value;
+			}
+
+			private int _virtualGet;
+			public virtual int VirtualGet
 			{
 				protected get
 				{
-					return virtualProperty;
+					return _virtualGet;
 				}
 				set
 				{
-					virtualProperty = value;
+					_virtualGet = value;
 				}
 
 			}
 
 			public int GetVirtual()
 			{
-				return Virtual;
+				return VirtualGet;
 			}
 		}
 
@@ -406,7 +523,9 @@ namespace Moq.Tests
 			int GetSomethingImpl();
 			void NonExistentMethod();
 			INested Nested { get; set; }
-			int Virtual { get; set; }
+			int this[int i, string s] { get; set; }
+			int VirtualGet { get; set; }
+			int VirtualSet { get; set; }
 		}
 
 		public abstract class MessageHandlerBase
@@ -425,6 +544,12 @@ namespace Moq.Tests
 		public interface MessageHandlerBaseish
 		{
 			void HandleImpl<TMessage>(TMessage message);
+		}
+
+		public class ExpectedException : Exception
+		{
+			private static ExpectedException expectedException = new ExpectedException();
+			public static ExpectedException Instance => expectedException;
 		}
 	}
 }